Reorganize upa commands
