--- upa/src/tool/genStandardJsonInput.ts
@@ -1,35 +1,0 @@
-import { command, option, positional, string } from "cmd-ts";
-import * as fs from "fs";
-
-export const genstandardjsoninput = command({
-  name: "gen-stanadard-json-input",
-  args: {
-    jsonArtifact: positional({
-      type: string,
-      displayName: "json-artifact",
-      description:
-        "JSON file generated by Hardhat in artifacts/build-info directory.",
-    }),
-    standardJsonInput: option({
-      type: string,
-      long: "standard-json-input-file",
-      short: "s",
-      description: "File to write standard json input to",
-      defaultValue: () => "standard-input.json",
-    }),
-  },
-  description: "Generate JSON json file for UPA contract instance",
-  handler: async function ({ jsonArtifact, standardJsonInput }): Promise<void> {
-    const inputJson = JSON.parse(fs.readFileSync(jsonArtifact, "ascii"))[
-      "input"
-    ];
-
-    const libraryObj = {};
-
-    const settingsJson = Object.assign(inputJson["settings"], libraryObj);
-
-    inputJson["settings"] = settingsJson;
-
-    fs.writeFileSync(standardJsonInput, JSON.stringify(inputJson));
-  },
-});
