+++ upa/src/tool/version.ts
@@ -4,7 +4,7 @@ import { command } from "cmd-ts";
 export const version = command({
   name: "version",
   args: {},
-  description: "Print the version and exit",
+  description: "Print the SDK version and exit",
   handler: () => {
     console.log(pkg.version);
   },
